+++ CHANGELOG.md
@@ -4,7 +4,7 @@ based on [Semantic Versioning](http://semver.org/)
 
 *1.0.4* 2015-02-14
 
-+ add reaction to ADD and REMOVE actions
++ add reaction to ADD, UPDATE, REMOVE and CLEAR actions
 
 *1.0.3* 2014-02-19